refactor : Member 도메인 작업에 따른 코드 수정
