+++ src/main/java/org/socialculture/platform/ticket/service/TicketServiceImpl.java
@@ -4,10 +4,9 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.socialculture.platform.global.apiResponse.exception.ErrorStatus;
 import org.socialculture.platform.global.apiResponse.exception.GeneralException;
+import org.socialculture.platform.member.repository.MemberRepository;
 import org.socialculture.platform.ticket.dto.response.TicketResponseDto;
-import org.socialculture.platform.ticket.entity.MemberEntity;
 import org.socialculture.platform.ticket.entity.TicketEntity;
-import org.socialculture.platform.ticket.repository.MemberRepository;
 import org.socialculture.platform.ticket.repository.TicketRepository;
 import org.springframework.dao.DataAccessException;
 import org.springframework.data.domain.PageRequest;
