+++ src/test/java/org/socialculture/platform/ticket/service/TicketServiceTest.java
@@ -9,11 +9,11 @@ import org.mockito.Mock;
 import org.mockito.MockitoAnnotations;
 import org.socialculture.platform.global.apiResponse.exception.ErrorStatus;
 import org.socialculture.platform.global.apiResponse.exception.GeneralException;
+import org.socialculture.platform.member.entity.MemberEntity;
+import org.socialculture.platform.member.repository.MemberRepository;
 import org.socialculture.platform.ticket.dto.response.TicketResponseDto;
-import org.socialculture.platform.ticket.entity.MemberEntity;
 import org.socialculture.platform.ticket.entity.PerformanceEntity;
 import org.socialculture.platform.ticket.entity.TicketEntity;
-import org.socialculture.platform.ticket.repository.MemberRepository;
 import org.socialculture.platform.ticket.repository.TicketRepository;
 import org.springframework.dao.DataAccessException;
